Add check entitlement on register
